Refactor components
