+++ lib/domino.rb
@@ -43,6 +43,9 @@ class Domino
   include Capybara::DSL
   extend  Capybara::DSL
 
+  require 'domino/attribute'
+  require 'domino/form'
+
   # Namespaced Domino::Error
   class Error < StandardError; end
 
@@ -186,170 +189,10 @@ class Domino
     end
   end
 
-  class Form < Domino
-    class Field
-      attr_reader :name, :locator, :options, :callback
-
-      def initialize(name, locator, options = {}, &callback)
-        @name = name
-        @locator = locator
-        @options = options
-        @callback = callback
-      end
-
-      def value(node)
-        val = read(node)
-        if val && callback.is_a?(Proc)
-          callback.call(val)
-        else
-          val
-        end
-      end
-
-      def read(node)
-        node.find_field(locator, options).value
-      end
-
-      def write(node, value)
-        node.fill_in(locator, with: value, **options)
-      end
-    end
-
-    class SelectField < Field
-      def write(node, value)
-        node.select value, from: locator
-      end
-    end
-
-    class BooleanField < Domino::Form::Field
-      def read(node)
-        node.find_field(locator, options).checked?
-      end
-
-      def write(node, value)
-        if value
-          node.check(locator, options)
-        else
-          node.uncheck(locator, options)
-        end
-      end
-    end
-
-    FIELD_TYPES = {
-      select: SelectField,
-      boolean: BooleanField
-    }.freeze
-
-    def self.key(k)
-      @key = k
-    end
-
-    def self.fields
-      @fields ||= {}
-    end
-
-    def self.field(*args, &callback)
-      options = args.last.is_a?(::Hash) ? args.pop : {}
-      attribute, locator = *args
-
-      locator ||= !@key.to_s.empty? ? "#{@key}[#{attribute}]" : attribute
-
-      field_type = options.delete(:as)
-      field_class = field_type.is_a?(Class) && field_type.ancestors.include?(Field) ? field_type : FIELD_TYPES[field_type] || Field
-
-      fields[attribute] = field_class.new(attribute, locator, options, &callback)
-
-      define_method :"#{attribute}" do
-        self.class.fields[attribute].value(node)
-      end
-
-      define_method :"#{attribute}=" do |value|
-        self.class.fields[attribute].write(node, value)
-      end
-    end
-
-    def self.create(attributes = {})
-      first.create(attributes)
-    end
-
-    def self.update(attributes = {})
-      first.update(attributes)
-    end
-
-    def create(attributes = {})
-      set(attributes)
-      save
-    end
-
-    def update(attributes = {})
-      set(attributes)
-      save
-    end
-
-    def set(attributes = {})
-      attributes.each { |k, v| send("#{k}=", v) }
-    end
-
-    def save
-      find('input[name="commit"]').click
-    end
-  end
-
   private
 
   # Store the capybara node internally
   def initialize(node)
     @node = node
-  end
-
-  class Attribute
-    attr_reader :name, :selector, :callback
-
-    def initialize(name, selector = nil, &callback)
-      @callback = callback
-      @name = name
-      @selector = selector || %(.#{name.to_s.tr('_', '-')})
-    end
-
-    def value(node)
-      val = value_before_typecast(node)
-
-      if val && callback.is_a?(Proc)
-        callback.call(val)
-      else
-        val
-      end
-    end
-
-    # Get the text of the first dom element matching a selector
-    #
-    #   Dom::Post.all.first.attribute('.title')
-    def value_before_typecast(node)
-      if combinator?
-        node[node_attribute_key] || node.matches_css?(combinator)
-      else
-        node.find(selector).text
-      end
-    rescue Capybara::ElementNotFound
-      nil
-    end
-
-    def match_value?(node, value)
-      value === value(node)
-    end
-
-    private
-
-    def combinator?
-      selector[0] == '&'.freeze
-    end
-
-    def combinator
-      @combinator ||= selector.sub(/&/, '') if combinator?
-    end
-
-    def node_attribute_key
-      @node_attribute_key ||= combinator.match(/(?<=\[).+?(?=\])/) { |m| m[0] }
-    end
   end
 end
