+++ lib/domino/form/select_field.rb
@@ -1,0 +1,5 @@
+class Domino::Form::SelectField < Domino::Form::Field
+  def write(node, value)
+    node.select value, from: locator
+  end
+end
